Tic Tac Toe app built using Ruby
This application should amount to the Tic Tac Toe children's game that can be played on a console. It is built using Ruby.
Currently, it contains two files, one for a main file that is used to run the app and an IntroductionMessages file that will display a welcome message when the app is executed.