Added Player Type to define a Human vs a Computer
PlayerType is an object that uses two symbols to identify the two options - Human, Computer. Symbols was used refer to the same string value instead of instantiating it every time it is called. It doesn't seem immutable though so I may need to use a freeze command or safe guard it another way to ensure that it doesn't change values.