Create Player object + basic tests
Created a Player that contains a type and a symbol that represents the player. Created associated unit test to test out the initialization message to check that the expected input is stored in the given variable. This was to verify that I wrote the right syntax and the test verified the right thing (when I change the expected value to a different one, the test failed as expected).