Added another test
Added another test for space character - updated logic to prevent user from using this as a symbol.
Also note: the reason for the logic change to use a regular expression (compared to the original C# logic) is that Ruby doesn't have a notion of "is_integer" to correctly identify one, the way C# does.