Added MatchTypeSetup
Added class to prompt the user for input on configuring the match type used during game setup.
"tc_test.rb" throughout this project is to test the scripts and print out the data to the terminal to test out those methods that write messages to the console. This way I can isolate the testing to those specific areas to make sure formatting, etc looks good (instead of having to run through the whole game).