Started working on BoardScoreValidator
Added validator for score of overrall board and for a given line on the board. So far, it basically has stable code for the raise ArgumentError lines.
Unit tests for raise assertions added as well.

Added code to calculate the score of a given line using a case statement - this requires tests to be written, next.